remove requiest ip
request-ip library doesn't do much, and looking through the last change
to address issues with 3rd party deps, the copy pasted code is both
inneficient and likely contains logical bugs that could throw at runtime
(eg
https://github.com/pbojinov/request-ip/blob/e1d0f4b89edf26c77cf62b5ef662ba1a0bd1c9fd/lib/is.js#L19
here erroneously checks if the value exists only when testing ipv4, then
might run regex match on undefined, causing a runtime error.

The package is a liability, and the benefit of adding it is minimal for
the overhead, even after the dependence on the other buggy library was
removed.

We can simplify the logic and remove another dependency on, in my
opinion, an unreliable library.